[#2] feat : filter 컴포넌트 mainpage 컴포넌트로 이동
