+++ src/pages/MainPage.jsx
@@ -1,11 +1,26 @@
 import Header from '../components/common/Header';
+import Filter from '../components/filter/Filter';
+import Place from '../components/filter/Place';
+import Price from '../components/filter/Price';
 import ProductList from './../components/product/ProductList';
 import { Container, Icon } from '@chakra-ui/react';
-import React from 'react';
+import React, { useState } from 'react';
 import { BsCart2 } from 'react-icons/bs';
 import { Link } from 'react-router-dom';
 
 export default function MainPage() {
+  const [currentTab, setCurrentTab] = useState('');
+  const [priceRange, setPriceRange] = useState([]);
+
+  const TABS = [
+    {
+      id: 1,
+      title: '가격순',
+      content: <Price setPriceRange={setPriceRange} />,
+    },
+    { id: 2, title: '공간순', content: <Place /> },
+  ];
+
   const rightComp = () => (
     <Link to='/reservations'>
       <Icon as={BsCart2} boxSize='6' />
@@ -15,6 +30,11 @@ export default function MainPage() {
   return (
     <Container maxW='container.sm' backgroundColor='white'>
       <Header rightComp={rightComp()} />
+      <Filter
+        TABS={TABS}
+        currentTab={currentTab}
+        setCurrentTab={setCurrentTab}
+      />
       <ProductList />
     </Container>
   );
